+++ changelog.md
@@ -12,6 +12,7 @@
 - Added `min_market_funds` and `max_market_funds` to `Product`
 - Made several explicit `Show` instances
 - Support list deposits endpoint
+- Support get deposit request
 
 # Version 0.8.2.0